requirements.txt
# requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
loguru==0.7.0
soundfile==0.12.1
librosa==0.10.0.post2
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.1
tqdm==4.65.0
opencv-python==4.7.0.72
Pillow==10.0.0
python-dotenv==1.0.0